refine order logic

+++ client/shop/service/cart.js
@@ -6,7 +6,7 @@ export const getCart = async (id) => {
   if (process.env.DEV_ENV !== "production")
     return generateDummyCart(Math.round(Math.random() * 10) + 1);
   try {
-    const response = await fetch(`${process.env.APP_URL}/product/cart/${id}`);
+    const response = await fetch(`${process.env.APP_URL}/products/cart/${id}`);
     if (response.ok) {
       const data = await response.json();
       return data;
@@ -23,7 +23,7 @@ export const addCartItem = async (id,product_id,quantity) => {
    if (process.env.DEV_ENV !== "production") return true;
    try {
     const response = await fetch(
-      `${process.env.APP_URL}/product/cart/${id}`,
+      `${process.env.APP_URL}/products/cart/${id}`,
       {
         method: "POST",
         headers: {
@@ -68,7 +68,7 @@ export const updateCartItem = async (id, product_id,quantity) => {
 };
 
 export const deleteAllCartItem = async (id) => {
-  // if (process.env.DEV_ENV !== "production") return true;
+  if (process.env.DEV_ENV !== "production") return true;
   try {
     const response = await fetch(
       `${process.env.APP_URL}/products/cart/${id}`,
@@ -88,10 +88,10 @@ export const deleteAllCartItem = async (id) => {
 };
 
 export const deleteCartItem = async (id, product_id) => {
-  // if (process.env.DEV_ENV !== "production") return true;
+  if (process.env.DEV_ENV !== "production") return true;
   try {
     const response = await fetch(
-      `${process.env.APP_URL}/products/carts/${id}/${product_id}`,
+      `${process.env.APP_URL}/products/cart/${id}/${product_id}`,
       {
         method: "DELETE",
       }
